refactor: extract common code & clean up
